pyproject.toml
[project]
name = "palette-transfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=9.0.0",
    "numpy>=1.20.0",
    "face-recognition>=1.3.0",
    "scipy>=1.7.0",
    "scikit-image>=0.19.0",
    "pytest>=8.4.1",
    "setuptools>=80.9.0",
    "scikit-learn>=1.7.1",
    "dask>=2025.7.0",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.0.7",
    "flask-wtf>=1.2.1",
    "fastmcp>=2.13.3",
    "uvicorn>=0.38.0",
    "pydantic>=2.12.5",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]
